+++ src/tree_grid/model_py.rs
@@ -1,4 +1,4 @@
-use core::f32;
+use core::f64;
 
 use super::{
     fitter::{RefineCandidate, TreeGridFitter},
@@ -13,9 +13,9 @@ use rand::prelude::*;
 #[pyclass(name = "RTGrid")]
 pub struct TreeGridPy {
     is_fitted: bool,
-    splits: Vec<Vec<f32>>,
-    intervals: Vec<Vec<(f32, f32)>>,
-    grid_values: Vec<Vec<f32>>,
+    splits: Vec<Vec<f64>>,
+    intervals: Vec<Vec<(f64, f64)>>,
+    grid_values: Vec<Vec<f64>>,
     hyperparameters: TreeGridParams,
 }
 
@@ -23,17 +23,17 @@ pub struct TreeGridPy {
 #[pyclass(name = "FResult")]
 pub struct FitResultPy {
     #[pyo3(get)]
-    pub err: f32,
+    pub err: f64,
     #[pyo3(get)]
-    pub residuals: Py<PyArray1<f32>>,
+    pub residuals: Py<PyArray1<f64>>,
     #[pyo3(get)]
-    pub y_hat: Py<PyArray1<f32>>,
+    pub y_hat: Py<PyArray1<f64>>,
 }
 
 #[pymethods]
 impl TreeGridPy {
     #[new]
-    fn new(n_iter: usize, split_try: usize, colsample_bytree: f32) -> Self {
+    fn new(n_iter: usize, split_try: usize, colsample_bytree: f64) -> Self {
         let splits = vec![];
         let intervals = vec![];
         let grid_values = vec![];
@@ -54,8 +54,8 @@ impl TreeGridPy {
     pub fn predict<'py>(
         &self,
         py: Python<'py>,
-        x: PyReadonlyArray2<'py, f32>,
-    ) -> PyResult<Bound<'py, PyArray1<f32>>> {
+        x: PyReadonlyArray2<'py, f64>,
+    ) -> PyResult<Bound<'py, PyArray1<f64>>> {
         let x = x.as_array();
         let mut y_hat = Array1::zeros(x.nrows());
         for (i, row) in x.axis_iter(Axis(0)).enumerate() {
@@ -75,8 +75,8 @@ impl TreeGridPy {
     pub fn fit<'py>(
         &mut self,
         py: Python<'py>,
-        x: PyReadonlyArray2<'py, f32>,
-        y: PyReadonlyArray1<'py, f32>,
+        x: PyReadonlyArray2<'py, f64>,
+        y: PyReadonlyArray1<'py, f64>,
     ) -> PyResult<FitResultPy> {
         let x = x.as_array();
         let y = y.as_array();
@@ -87,7 +87,7 @@ impl TreeGridPy {
         for _ in 0..self.hyperparameters.n_iter {
             // sample random columns to split on
             let n_cols_to_sample =
-                (self.hyperparameters.colsample_bytree * x.ncols() as f32) as usize;
+                (self.hyperparameters.colsample_bytree * x.ncols() as f64) as usize;
 
             let split_idx: Vec<usize> = (0..self.hyperparameters.split_try)
                 .map(|_| rng.gen_range(0..x.nrows()))
@@ -99,7 +99,7 @@ impl TreeGridPy {
             let col_idx = possible_indices[0..n_cols_to_sample].to_vec();
 
             let mut best_candidate: Option<RefineCandidate> = None;
-            let mut best_err_diff = f32::NEG_INFINITY;
+            let mut best_err_diff = f64::NEG_INFINITY;
             for col in &col_idx {
                 for idx in &split_idx {
                     let split = x[[*idx, *col]];
